+++ src/utils/go_to_page.ts
@@ -1,3 +1,4 @@
 export const HOME = '/';
 export const BLOCKS = '/blocks';
+export const BLOCK_DETAILS = (hash: string) => `/blocks/${hash}`;
 export const TRANSACTIONS = '/transactions';